+++ autopush/endpoint.py
@@ -1010,7 +1010,7 @@ class RegistrationHandler(AutoendpointHandler):
         self.chid = params["channelID"]
         if new_uaid:
             d = Deferred()
-            d.addCallback(router.register, params.get("data", {}))
+            d.addCallback(router.register, params)
             d.addCallback(self._save_router_data, router_type)
             d.addCallback(self._make_endpoint)
             d.addCallback(self._return_endpoint, new_uaid, router)